move tests proper into `tests/type_checker`
