+++ components/lark-test/src/lib.rs
@@ -6,11 +6,11 @@ use lark_query_system::LarkDatabase;
 use parser::HasParserState;
 use parser::HasReaderState;
 
-trait ErrorSpec {
+pub trait ErrorSpec {
     fn check_errors(&self, errors: &[RangedDiagnostic]);
 }
 
-struct NoErrors;
+pub struct NoErrors;
 
 impl ErrorSpec for NoErrors {
     fn check_errors(&self, errors: &[RangedDiagnostic]) {
@@ -55,7 +55,7 @@ impl ErrorSpec for &str {
     }
 }
 
-fn run_test(text: &str, error_spec: impl ErrorSpec) {
+pub fn run_test(text: &str, error_spec: impl ErrorSpec) {
     let mut db = LarkDatabase::default();
     let path1_str = "path1";
     let path1_interned = db.intern_string("path1");
@@ -81,53 +81,4 @@ fn run_test(text: &str, error_spec: impl ErrorSpec) {
             panic!("cancelled?!");
         }
     }
-}
-
-#[test]
-fn bad_identifier() {
-    run_test(
-        "def new(msg: bool,) -> bool { msg1 }",
-        "                              ~~~~",
-    );
-}
-
-#[test]
-fn bad_callee() {
-    run_test(
-        "def foo(msg: bool,) -> bool { bar(msg) }",
-        "                              ~~~",
-    );
-}
-
-#[test]
-fn correct_call() {
-    run_test(
-        "def foo(msg: bool,) { bar(msg) } def bar(arg:bool,) { }",
-        NoErrors,
-    );
-}
-
-#[test]
-fn wrong_num_of_arguments() {
-    run_test(
-        "def foo(msg: bool,) -> bool { bar(msg) } def bar(arg:bool, arg2:bool) { }",
-        "                              ~~~~~~~~",
-    );
-}
-
-#[test]
-fn wrong_return_type() {
-    // `bar` returns unit, we expect `bool`
-    run_test(
-        "def foo(msg: bool,) -> bool { bar(msg) } def bar(arg:bool,) { }",
-        "                              ~~~~~~~~",
-    );
-}
-
-#[test]
-fn wrong_type_of_arguments() {
-    run_test(
-        "def foo(msg: int,) { bar(msg) } def bar(arg:bool,) { }",
-        "                         ~~~",
-    );
 }
